Back to a single class to observe av-off files & folders
