+++ src/com/owncloud/android/services/observer/AvailableOfflineObserver.java
@@ -2,7 +2,7 @@
  *   ownCloud Android client application
  *
  *   @author David A. Velasco
- *   Copyright (C) 2016 ownCloud Inc.
+ *   Copyright (C) 2016 ownCloud GmbH.
  *
  *   This program is free software: you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License version 2,
@@ -20,43 +20,206 @@
 
 package com.owncloud.android.services.observer;
 
+import android.accounts.Account;
+import android.content.Context;
+import android.content.Intent;
+import android.os.FileObserver;
+
+import com.owncloud.android.datamodel.FileDataStorageManager;
+import com.owncloud.android.datamodel.OCFile;
+import com.owncloud.android.lib.common.operations.RemoteOperationResult;
+import com.owncloud.android.lib.common.operations.RemoteOperationResult.ResultCode;
+import com.owncloud.android.lib.common.utils.Log_OC;
+import com.owncloud.android.operations.SynchronizeFileOperation;
+import com.owncloud.android.ui.activity.ConflictsResolveActivity;
+
+import java.io.File;
+import java.util.HashMap;
+import java.util.Map;
+
 /**
- * Interface for both recursive and non recursive folder observers used by {@link FileObserverService}
+ * Observer watching a folder to request the synchronization of kept-in-sync files
+ * inside it.
+ * 
+ * Takes into account two possible update cases:
+ *  - an editor directly updates the file;
+ *  - an editor works on a temporary file, and later replaces the kept-in-sync file with the
+ *  former.
+ *  
+ *  The second case requires to monitor the folder parent of the files, since a direct 
+ *  {@link FileObserver} on it will not receive more events after the file is deleted to
+ *  be replaced.
  */
-public interface AvailableOfflineObserver {
+public class AvailableOfflineObserver {
 
+    private static String TAG = AvailableOfflineObserver.class.getSimpleName();
+
+    private static int UPDATE_MASK = (
+            FileObserver.ATTRIB | FileObserver.MODIFY |
+            FileObserver.MOVED_TO | FileObserver.CLOSE_WRITE
+    );
+
+    private static int IN_IGNORE = 32768;
+    /*
+    private static int ALL_EVENTS_EVEN_THOSE_NOT_DOCUMENTED = 0x7fffffff;   // NEVER use 0xffffffff
+    */
+
+    private String mPath;
+    private Account mAccount;
+    private Context mContext;
+    private Map<String, Boolean> mObservedChildren;
+    private FileObserver mWrappedObserver;
 
     /**
-     * Starts the observation
+     * Constructor.
+     *
+     * Initializes the observer to receive events about the update of the passed folder, and
+     * its children files.
+     *
+     * @param path          Absolute path to the local folder to watch.
+     * @param account       OwnCloud account associated to the folder.
+     * @param context       Used to start an operation to synchronize the file, when needed.
      */
-    void startWatching();
+    public AvailableOfflineObserver(String path, Account account, Context context) {
+        if (path == null)
+            throw new IllegalArgumentException("NULL path argument received");
+        if (account == null)
+            throw new IllegalArgumentException("NULL account argument received");
+        if (context == null)
+            throw new IllegalArgumentException("NULL context argument received");
+        
+        mPath = path;
+        mAccount = account;
+        mContext = context;
+        mObservedChildren = new HashMap<String, Boolean>();
 
+        //mWrappedObserver = new FileObserver(mPath, UPDATE_MASK) {
+        mWrappedObserver = new RecursiveFileObserver(mPath, UPDATE_MASK) {
 
-    /**
-     * Stops the observation
-     */
-    public void stopWatching();
+            /**
+             * Receives and processes events about updates of the monitored folder and its children files.
+             *
+             * @param event     Kind of event occurred.
+             * @param path      Relative path of the file referred by the event.
+             */
+            @Override
+            public void onEvent(int event, String path) {
+                Log_OC.d(TAG, "Got event " + event + " on FOLDER " + mPath + " about "
+                    + ((path != null) ? path : ""));
 
+                boolean shouldSynchronize = false;
+                synchronized(mObservedChildren) {
+                    if (path != null && path.length() > 0 && mObservedChildren.containsKey(path)) {
+
+                        if (    ((event & FileObserver.MODIFY) != 0) ||
+                            ((event & FileObserver.ATTRIB) != 0) ||
+                            ((event & FileObserver.MOVED_TO) != 0) ) {
+
+                            if (mObservedChildren.get(path) != true) {
+                                mObservedChildren.put(path, Boolean.valueOf(true));
+                            }
+                        }
+
+                        if ((event & FileObserver.CLOSE_WRITE) != 0 && mObservedChildren.get(path)) {
+                            mObservedChildren.put(path, Boolean.valueOf(false));
+                            shouldSynchronize = true;
+                        }
+                    }
+                }
+                if (shouldSynchronize) {
+                    startSyncOperation(path);
+                }
+
+                if ((event & IN_IGNORE) != 0 &&
+                    (path == null || path.length() == 0)) {
+                    Log_OC.d(TAG, "Stopping the observance on " + mPath);
+                }
+
+            }
+        };
+    }
+
+    public void stopWatching() {
+        mWrappedObserver.stopWatching();
+    }
 
     /**
      * Adds a child file to the list of files observed by the folder observer.
      * 
-     * @param relativePath      Relative path to a file inside the observed folder.
+     * @param fileName         Name of a file inside the observed folder. 
      */
-    void startWatching(String relativePath);
+    public void startWatching(String fileName) {
+        synchronized (mObservedChildren) {
+            if (!mObservedChildren.containsKey(fileName)) {
+                mObservedChildren.put(fileName, Boolean.valueOf(false));
+            }
+        }
+        
+        if (new File(mPath).exists()) {
+            mWrappedObserver.startWatching();
+            Log_OC.d(TAG, "Started watching parent folder " + mPath + "/");
+        }
+        // else - the observance can't be started on a file not existing;
+    }
 
     
     /**
      * Removes a child file from the list of files observed by the folder observer.
      * 
-     * @param relativePath      Relative path to a file inside the observed folder.
+     * @param relativePath         Name of a file inside the observed folder.
      */
-    void stopWatching(String relativePath);
-
+    public void stopWatching(String relativePath) {
+        synchronized (mObservedChildren) {
+            mObservedChildren.remove(relativePath);
+            if (mObservedChildren.isEmpty()) {
+                mWrappedObserver.stopWatching();
+                Log_OC.d(TAG, "Stopped watching parent folder " + mPath + "/");
+            }
+        }
+    }
 
     /**
      * @return      'True' when the folder is not watching any file inside.
      */
-    boolean isEmpty();
+    public boolean isEmpty() {
+        synchronized (mObservedChildren) {
+            return mObservedChildren.isEmpty();
+        }
+    }
+
+    public void startWatching() {
+        mWrappedObserver.startWatching();
+    }
+
+
+    /**
+     * Triggers an operation to synchronize the contents of a file inside the observed folder with
+     * its remote counterpart in the associated ownCloud account.
+     *    
+     * @param fileName          Name of a file inside the watched folder.
+     */
+    private void startSyncOperation(String fileName) {
+        FileDataStorageManager storageManager = 
+                new FileDataStorageManager(mAccount, mContext.getContentResolver());
+        // a fresh object is needed; many things could have occurred to the file
+        // since it was registered to observe again, assuming that local files
+        // are linked to a remote file AT MOST, SOMETHING TO BE DONE;
+        OCFile file = storageManager.getFileByLocalPath(mPath + File.separator + fileName);
+        SynchronizeFileOperation sfo = 
+                new SynchronizeFileOperation(file, null, mAccount, true, mContext);
+        RemoteOperationResult result = sfo.execute(storageManager, mContext);
+        if (result.getCode() == ResultCode.SYNC_CONFLICT) {
+            // ISSUE 5: if the user is not running the app (this is a service!),
+            // this can be very intrusive; a notification should be preferred
+            Intent i = new Intent(mContext, ConflictsResolveActivity.class);
+            i.setFlags(i.getFlags() | Intent.FLAG_ACTIVITY_NEW_TASK);
+            i.putExtra(ConflictsResolveActivity.EXTRA_FILE, file);
+            i.putExtra(ConflictsResolveActivity.EXTRA_ACCOUNT, mAccount);
+            mContext.startActivity(i);
+        }
+        // TODO save other errors in some point where the user can inspect them later;
+        // or maybe just toast them;
+        // or nothing, very strange fails
+    }
 
 }
